ResolvedLicenseInfo: Properly filterExcluded() detected licenses
If a resolved license info only corresponds to the license source
"detected", then the license info is properly filtered out by the
already existing filter condition in filterExcluded(). However,
if a resolved license info corresponds to the license source "detected"
and to at least one more arbitrary license source, then the original
expressions for detected licenses are not filtered according to the path
excludes. In the following example the detected license won't get filtered
out no matter if they are excluded:

declared license: A
detected license: A OR B

So, B is accidentally included into the effective license which is
visible from the static HTML report as well as in the notice file. Fix
that by filtering the `originalExpressions` by their exluded flag.

Signed-off-by: Frank Viernau <frank.viernau@here.com>